fix : signUp 로직 패키지 변경(auth -> member), 정적팩토리메서드 삭제
