--- src/main/java/com/finalteam4/danggeunplanner/auth/dto/SignUpRequestDto.java
@@ -1,10 +1,0 @@
-package com.finalteam4.danggeunplanner.auth.dto;
-
-import lombok.Getter;
-
-@Getter
-public class SignUpRequestDto {
-    private String email;
-    private String password;
-
-}
