+++ src/main/java/com/finalteam4/danggeunplanner/member/controller/MemberController.java
@@ -1,7 +1,13 @@
 package com.finalteam4.danggeunplanner.member.controller;
 
+import com.finalteam4.danggeunplanner.common.response.ResponseMessage;
+import com.finalteam4.danggeunplanner.member.dto.request.SignUpRequestDto;
 import com.finalteam4.danggeunplanner.member.service.MemberService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -10,4 +16,12 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/member")
 public class MemberController {
     private final MemberService memberService;
+
+    @PostMapping("/signup")
+    public ResponseEntity<?> signUp(@RequestBody SignUpRequestDto signUpRequestDto){
+
+        memberService.signUp(signUpRequestDto);
+
+        return new ResponseEntity<>(new ResponseMessage("회원가입 성공",null), HttpStatus.CREATED);
+    }
 }
