+++ src/main/java/com/finalteam4/danggeunplanner/member/dto/request/SignUpRequestDto.java
@@ -1,0 +1,11 @@
+package com.finalteam4.danggeunplanner.member.dto.request;
+
+import lombok.Getter;
+
+@Getter
+public class SignUpRequestDto {
+
+        private String email;
+        private String password;
+
+}
