+++ src/main/java/com/finalteam4/danggeunplanner/member/entity/Member.java
@@ -39,8 +39,4 @@ public class Member {
         this.password = password;
         this.profileImage = "https://item.kakaocdn.net/do/1e917e59f980468a78f2bff7dcc25ac215b3f4e3c2033bfd702a321ec6eda72c";
     }
-    public static Member signUpMember(String email, String password){
-        return new Member(email, password);
-
-    }
 }
