+++ src/main/java/com/finalteam4/danggeunplanner/member/service/MemberService.java
@@ -1,13 +1,32 @@
 package com.finalteam4.danggeunplanner.member.service;
 
+import com.finalteam4.danggeunplanner.common.exception.DanggeunPlannerException;
+import com.finalteam4.danggeunplanner.member.dto.request.SignUpRequestDto;
+import com.finalteam4.danggeunplanner.member.entity.Member;
 import com.finalteam4.danggeunplanner.member.repository.MemberRepository;
+import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import static com.finalteam4.danggeunplanner.common.exception.ErrorCode.DUPLICATED_EMAIL;
 
 @Service
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
 public class MemberService {
     private final MemberRepository memberRepository;
+
+    @Transactional
+    public void signUp(SignUpRequestDto signUpRequestDto) {
+        String email = signUpRequestDto.getEmail();
+        String password = signUpRequestDto.getPassword();
+
+        memberRepository.findByEmail(email)
+                .ifPresent( m -> {
+                    throw new DanggeunPlannerException(DUPLICATED_EMAIL);
+                });
+
+        memberRepository.save(new Member(email, password));
+    }
 }
